address small comments
