+++ .github/workflows/templates-build-push-image.yml
@@ -1,6 +1,5 @@
-name: Build and Push (Linux)
 on:
-  workflow_dispatch:
+  workflow_call:
     inputs:
       image_name:
         required: true
@@ -16,8 +15,6 @@ jobs:
   linux:
     name: Build & Push (Linux)
     runs-on: ubuntu-latest
-    permissions:
-      packages: write
     steps:
       - name: Checkout Code
         uses: actions/checkout@v4
@@ -43,7 +40,7 @@ jobs:
         uses: docker/login-action@v3
         with:
           registry: ghcr.io
-          username: hkfgo
+          username: tomkerkhove
           password: ${{ secrets.GITHUB_TOKEN }}
 
       - name: Build and push preview image
@@ -52,5 +49,5 @@ jobs:
           build-args: VERSION="${{ env.artifact_full_version }}"
           context: ./src/
           file: ./src/${{ inputs.project_name }}/Dockerfile.linux
-          tags: ghcr.io/hkfgo/${{ env.image_commit_uri }},ghcr.io/hkfgo/${{ env.image_latest_uri }}
+          tags: ${{ env.image_commit_uri }},${{ env.image_latest_uri }}
           push: true
